4.1.7 Вынесет из utils функции для работы с фильмом в отдельный файл
